fixed data type of aGetAbsPosition function

+++ Logisim(clone)/Entities/gates/socket.h
@@ -28,7 +28,7 @@ public:
     }
     ~Socket() { }
     
-    glm::ivec2 GetAbsPosition() const {
+    glm::vec2 GetAbsPosition() const {
         return this->m_absPosition;
     }
